Add CQ Logo in the Account Page for all Screens
